refactor(materials): align domain model and implement DTOs (Phase 20a)
- Make id required in Material and MaterialConsumption interfaces
- Apply standard parsePositiveInt validation in materials parsers
- Implement CreateMaterialRequest and UpdateMaterialRequest DTOs
- Implement CreateMaterialConsumptionRequest DTO
- Add JSON serialization helpers for all materials DTOs
- Refactor materials types into dedicated domain files and export via index
